+++ modules/ccv2/ui/src/sap/commerce/toolset/ccv2/ui/view/CCv2BuildsDataView.kt
@@ -31,9 +31,7 @@ import sap.commerce.toolset.HybrisIcons
 import sap.commerce.toolset.ccv2.CCv2UiConstants
 import sap.commerce.toolset.ccv2.dto.CCv2BuildDto
 import sap.commerce.toolset.ccv2.settings.state.CCv2Subscription
-import sap.commerce.toolset.ccv2.ui.CCv2ToolWindowContentTab
-import sap.commerce.toolset.ccv2.ui.date
-import sap.commerce.toolset.ccv2.ui.sUser
+import sap.commerce.toolset.ccv2.ui.*
 import sap.commerce.toolset.ui.actionsButton
 import sap.commerce.toolset.ui.copyLink
 import sap.commerce.toolset.ui.scrollPanel
@@ -110,17 +108,13 @@ object CCv2BuildsDataView : CCv2DataView<CCv2BuildDto>() {
 
             panel {
                 row {
-                    icon(HybrisIcons.CCv2.Build.BRANCH).gap(RightGap.SMALL)
-                    copyLink(project, "Branch", build.branch, "Build Branch copied to clipboard")
+                    branch(project, build)
                 }
             }.gap(RightGap.COLUMNS)
 
             panel {
                 row {
-                    icon(build.statusIcon)
-                        .gap(RightGap.SMALL)
-                    label(build.status.title)
-                        .comment("Status")
+                    status(build)
                 }
             }.gap(RightGap.COLUMNS)